+++ ja2lib/Strategic/MapScreen.h
@@ -23,7 +23,7 @@ void ReBuildCharactersList(void);
 BOOLEAN HandlePreloadOfMapGraphics(void);
 void HandleRemovalOfPreLoadedMapGraphics(void);
 
-void ChangeSelectedMapSector(INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+void ChangeSelectedMapSector(u8 sMapX, u8 sMapY, INT8 bMapZ);
 
 BOOLEAN CanToggleSelectedCharInventory(void);
